Open license link in a new tab
This commit updates the links pointing to the license file on the `info.html` page to improve user experience.

The `target="_blank"` attribute has been added to the `license.txt` anchor tags. This ensures that the license file now opens in a new browser tab, preventing the user from navigating away from the main web interface when viewing the license.